added xp-formula and max-level
